MultiMedia-LLM-Summarization-Tool/requirements.txt
PyPDF2
ollama
openai
openai-whisper
python-dotenv
jupyterlab
ipywidgets
streamlit
anthropic
google.generativeai